requirements.txt
numpy
python-magic